+++ gunc/tests/test_external_tools.py
@@ -1,5 +1,8 @@
 from ..external_tools import *
-from pkg_resources import resource_filename
+from importlib.resources import files as _pkg_files
+
+def resource_filename(package, resource):
+    return str(_pkg_files(package).joinpath(resource))
 
 fasta_file = resource_filename(__name__, "test_data/tiny_test.faa.gz")